canted_view_to_proportional_circumscribed_orthogonal to client_core
